+++ CHANGELOG.rst
@@ -78,6 +78,7 @@ Removed
 * `@HiromuHota`_: Remove __repr__ from each mixin class as the referenced attributes are not available.
 * `@HiromuHota`_: Remove the dependency on nltk, but ``PorterStemmer()`` can still be used,
   if it is provided as ``DictionaryMatch(stemmer=PorterStemmer())``.
+* `@HiromuHota`_: Remove ``_NgramMatcher`` and ``_FigureMatcher`` as they are no longer needed.
 
 Fixed
 ^^^^^